pyproject.toml
[tool.poetry]
name = "DJANGODRF"
version = "0.1.0"
description = ""
authors = ["lucasljoliveira-olist <lucasjoao.oliveira@olist.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.1.6"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
